backend/requirements.txt
torch
torchvision
fastapi
uvicorn
numpy
opencv-python-headless
matplotlib
scipy
segment_anything
python-multipart
pillow
geopy